% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base.R
\name{.as_comma}
\alias{.as_comma}
\title{Format numbers with comma as thousands separator}
\usage{
.as_comma(x, digits = 0L)
}
\arguments{
\item{x}{Integer or numeric vector.}

\item{digits}{Number of digits after the decimal point. Default \code{0}.}
}
\value{
Character vector of the same length as \code{x}.
}
\description{
Format numbers with comma as thousands separator
}
\keyword{internal}
