% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base.R
\name{.assert_length}
\alias{.assert_length}
\title{Assert that \code{x} has nonzero length}
\usage{
.assert_length(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Invisibly returns \code{x} on success; otherwise aborts.
}
\description{
Assert that \code{x} has nonzero length
}
\keyword{internal}
