% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.element_axis_text}
\alias{.element_axis_text}
\title{Axis text element with size=0 sentinel for element_blank}
\usage{
.element_axis_text(size, face, angle, hjust, vjust, family)
}
\description{
Axis text element with size=0 sentinel for element_blank
}
\keyword{internal}
