% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.element_grid}
\alias{.element_grid}
\title{Grid line element with a boolean show toggle}
\usage{
.element_grid(show, color, linewidth)
}
\description{
Grid line element with a boolean show toggle
}
\keyword{internal}
