% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.element_ticks}
\alias{.element_ticks}
\title{Axis ticks element with size=0 sentinel for element_blank}
\usage{
.element_ticks(size)
}
\description{
Axis ticks element with size=0 sentinel for element_blank
}
\keyword{internal}
