% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\name{.find_maturity}
\alias{.find_maturity}
\title{Internal: locate the first mature ata link from an \code{ATASummary}}
\usage{
.find_maturity(
  x,
  cv_threshold = 0.1,
  rse_threshold = 0.05,
  min_valid_ratio = 0.5,
  min_n_valid = 3L,
  min_run = 1L
)
}
\description{
Internal: locate the first mature ata link from an \code{ATASummary}
}
\keyword{internal}
