% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base.R
\name{.has_cols}
\alias{.has_cols}
\title{Whether all of \code{cols} are present in \code{names(df)}}
\usage{
.has_cols(df, cols)
}
\arguments{
\item{df}{A data.frame (or similar with \code{names()}).}

\item{cols}{Character vector of column names.}
}
\value{
Logical scalar.
}
\description{
Whether all of \code{cols} are present in \code{names(df)}
}
\keyword{internal}
