% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{.mack_param_var}
\alias{.mack_param_var}
\title{Compute Mack parameter variance for a single cohort}
\usage{
.mack_param_var(target_proj, f_selected, f_var, last_obs)
}
\description{
Internal helper computing:

\deqn{
  \mathrm{param}_{i,k+1} =
    f_k^2 \cdot \mathrm{param}_{i,k} +
    \hat{C}_{i,k}^2 \cdot \mathrm{Var}(\hat{f}_k)
}
}
\keyword{internal}
