% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{.ratio_summary}
\alias{.ratio_summary}
\title{Summarise an \code{RatioFit} object by cohort}
\usage{
.ratio_summary(x)
}
\arguments{
\item{x}{An object of class \code{"RatioFit"}.}
}
\value{
The input object with \verb{$summary} set.
}
\description{
Summarise an \code{RatioFit} object by cohort
}
\keyword{internal}
