% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base.R
\name{.remove_class}
\alias{.remove_class}
\title{Remove class(es) from an object}
\usage{
.remove_class(x, classes)
}
\arguments{
\item{x}{An R object.}

\item{classes}{Character vector of class names to remove.}
}
\value{
\code{x} with class attribute updated.
}
\description{
Remove one or more class names from an object's class vector. For
\code{data.table} input, uses \code{\link[data.table:setattr]{data.table::setattr()}} to avoid copy.
}
\keyword{internal}
