% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base.R
\name{.update_class}
\alias{.update_class}
\title{Update class attribute: remove then prepend}
\usage{
.update_class(x, remove = NULL, prepend = NULL)
}
\arguments{
\item{x}{An R object.}

\item{remove}{Character vector of class names to remove. \code{NULL} to skip.}

\item{prepend}{Character vector of class names to prepend. \code{NULL} to skip.}
}
\value{
\code{x} with class attribute updated.
}
\description{
Convenience wrapper that first removes classes listed in \code{remove},
then prepends classes listed in \code{prepend}.
}
\keyword{internal}
