% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base.R
\name{longer}
\alias{longer}
\title{Reshape an object to long form (S3 generic)}
\usage{
longer(x, ...)
}
\arguments{
\item{x}{An object to reshape.}

\item{...}{Passed to methods.}
}
\value{
A long-form object as defined by the dispatched method.
}
\description{
Mirrors the S3 generic originally provided by the \code{instead} package.
Concrete methods (\code{longer.Triangle}, \code{longer.TriangleSummary}) dispatch on
domain-specific classes and typically return the pre-computed long-form
data stored in \code{attr(x, "longer")}.
}
