% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-vis.R
\name{plot.ATAFit}
\alias{plot.ATAFit}
\title{Plot an ata fit}
\usage{
\method{plot}{ATAFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ATAFit"}.}

\item{...}{Arguments passed to \code{\link[=plot.Link]{plot.Link()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"ATAFit"} by delegating to \code{\link[=plot.Link]{plot.Link()}}
on the underlying \code{Link} data stored in \code{x$link} with \code{model = "ata"}.
}
\seealso{
\code{\link[=plot.Link]{plot.Link()}}, \code{\link[=fit_ata]{fit_ata()}}
}
