% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-vis.R
\name{plot.BFFit}
\alias{plot.BFFit}
\title{Plot a Bornhuetter-Ferguson fit}
\usage{
\method{plot}{BFFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BFFit"}.}

\item{...}{Forwarded to \code{.plot_projection_fit()} -- see \code{\link[=plot.SAFit]{plot.SAFit()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Projection plot for a \code{"BFFit"} -- observed and projected cumulative
loss by cohort, delegated to \code{.plot_projection_fit()}.
}
