% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl-vis.R
\name{plot.CCFit}
\alias{plot.CCFit}
\title{Plot a Cape Cod fit}
\usage{
\method{plot}{CCFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"CCFit"}.}

\item{...}{Forwarded to \code{.plot_projection_fit()} -- see \code{\link[=plot.SAFit]{plot.SAFit()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Projection plot for a \code{"CCFit"} -- observed and projected cumulative
loss by cohort, delegated to \code{.plot_projection_fit()}.
}
