% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed-vis.R
\name{plot.ED}
\alias{plot.ED}
\title{Plot ED intensity diagnostics}
\usage{
\method{plot}{ED}(
  x,
  type = c("summary", "box", "point"),
  alpha = 1,
  scales = c("fixed", "free", "free_x", "free_y"),
  nrow = NULL,
  ncol = NULL,
  theme = c("view", "save", "shiny"),
  x.angle = 90,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"ED"}.}

\item{type}{One of \code{"summary"}, \code{"box"}, or \code{"point"}.}

\item{alpha}{Numeric scalar. Default is \code{1}.}

\item{scales}{Facet scale argument.}

\item{nrow, ncol}{Facet dimensions.}

\item{theme}{Theme string.}

\item{x.angle}{Numeric angle for x-axis tick labels. Default is \code{90}
to prevent overlap of the \code{from-to} link labels.}

\item{...}{Additional arguments passed to \code{\link[=.switch_theme]{.switch_theme()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise diagnostic summaries from an \code{"ED"} object. Internally
calls the \code{summary()} method on an \code{ED} object.
}
