% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-vis.R
\name{plot.EDFit}
\alias{plot.EDFit}
\title{Plot an ED fit}
\usage{
\method{plot}{EDFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"EDFit"}.}

\item{...}{Arguments passed to \code{\link[=plot.Link]{plot.Link()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"EDFit"} by delegating to \code{\link[=plot.Link]{plot.Link()}}
on the underlying \code{Link} data stored in \code{x$link} with \code{model = "ed"}.
}
\seealso{
\code{\link[=plot.Link]{plot.Link()}}, \code{\link[=fit_ed]{fit_ed()}}
}
