% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-vis.R
\name{plot.IntensityFit}
\alias{plot.IntensityFit}
\title{Plot an Intensity fit}
\usage{
\method{plot}{IntensityFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"IntensityFit"}.}

\item{...}{Arguments passed to \code{\link[=plot.Link]{plot.Link()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"IntensityFit"} by delegating to
\code{\link[=plot.Link]{plot.Link()}} on the underlying \code{Link} data stored in \code{x$link} with
\code{model = "ed"}. \code{IntensityFit} is the factor-level diagnostic for
the exposure-driven (ED) workflow and mirrors how \code{\link[=plot.ATAFit]{plot.ATAFit()}}
delegates for the multiplicative (CL) side.
}
\seealso{
\code{\link[=plot.Link]{plot.Link()}}, \code{\link[=fit_intensity]{fit_intensity()}}, \code{\link[=plot.ATAFit]{plot.ATAFit()}}
}
