% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-vis.R
\name{plot_triangle.ATAFit}
\alias{plot_triangle.ATAFit}
\title{Triangle heatmap for an ata fit}
\usage{
\method{plot_triangle}{ATAFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ATAFit"}.}

\item{...}{Arguments passed to \code{\link[=plot_triangle.Link]{plot_triangle.Link()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"ATAFit"} as a triangle-style heatmap
by delegating to \code{\link[=plot_triangle.Link]{plot_triangle.Link()}} on the underlying \code{Link} data
stored in \code{x$link} with \code{model = "ata"}.
}
\seealso{
\code{\link[=plot_triangle.Link]{plot_triangle.Link()}}, \code{\link[=fit_ata]{fit_ata()}}
}
