% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-vis.R
\name{plot_triangle.EDFit}
\alias{plot_triangle.EDFit}
\title{Triangle heatmap for an ED fit}
\usage{
\method{plot_triangle}{EDFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"EDFit"}.}

\item{...}{Arguments passed to \code{\link[=plot_triangle.Link]{plot_triangle.Link()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"EDFit"} as a triangle-style heatmap
by delegating to \code{\link[=plot_triangle.Link]{plot_triangle.Link()}} on the underlying \code{Link} data
stored in \code{x$link} with \code{model = "ed"}.
}
\seealso{
\code{\link[=plot_triangle.Link]{plot_triangle.Link()}}, \code{\link[=fit_ed]{fit_ed()}}
}
