% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-vis.R
\name{plot_triangle.IntensityFit}
\alias{plot_triangle.IntensityFit}
\title{Triangle heatmap for an Intensity fit}
\usage{
\method{plot_triangle}{IntensityFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"IntensityFit"}.}

\item{...}{Arguments passed to \code{\link[=plot_triangle.Link]{plot_triangle.Link()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"IntensityFit"} as a triangle-style
heatmap by delegating to \code{\link[=plot_triangle.Link]{plot_triangle.Link()}} on the underlying
\code{Link} data stored in \code{x$link} with \code{model = "ed"}.
}
\seealso{
\code{\link[=plot_triangle.Link]{plot_triangle.Link()}}, \code{\link[=fit_intensity]{fit_intensity()}},
\code{\link[=plot_triangle.ATAFit]{plot_triangle.ATAFit()}}
}
