% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.R
\name{print.BFFit}
\alias{print.BFFit}
\title{Print method for \code{BFFit}}
\usage{
\method{print}{BFFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BFFit"}.}

\item{...}{Unused.}
}
\description{
Print method for \code{BFFit}
}
