% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.BootstrapTriangle}
\alias{print.BootstrapTriangle}
\title{Print method for BootstrapTriangle}
\usage{
\method{print}{BootstrapTriangle}(x, ...)
}
\arguments{
\item{x}{A \code{BootstrapTriangle} object.}

\item{...}{Unused.}
}
\description{
Print method for BootstrapTriangle
}
