% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capecod.R
\name{print.CapeCodFit}
\alias{print.CapeCodFit}
\title{Print method for \code{CapeCodFit}}
\usage{
\method{print}{CapeCodFit}(x, ...)
}
\arguments{
\item{x}{A \code{CapeCodFit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{CapeCodFit}
}
