% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{print.CCFit}
\alias{print.CCFit}
\title{Print method for \code{CCFit}}
\usage{
\method{print}{CCFit}(x, ...)
}
\arguments{
\item{x}{A \code{CCFit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{CCFit}
}
