% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{print.ExposureFit}
\alias{print.ExposureFit}
\title{Print method for \code{ExposureFit}}
\usage{
\method{print}{ExposureFit}(x, ...)
}
\arguments{
\item{x}{An \code{ExposureFit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{ExposureFit}
}
