% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{print.LossFit}
\alias{print.LossFit}
\title{Print method for \code{LossFit}}
\usage{
\method{print}{LossFit}(x, ...)
}
\arguments{
\item{x}{A \code{LossFit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{LossFit}
}
