% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prem.R
\name{print.PremFit}
\alias{print.PremFit}
\title{Print method for \code{PremFit}}
\usage{
\method{print}{PremFit}(x, ...)
}
\arguments{
\item{x}{A \code{PremFit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{PremFit}
}
