% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premium.R
\name{print.PremiumFit}
\alias{print.PremiumFit}
\title{Print method for \code{PremiumFit}}
\usage{
\method{print}{PremiumFit}(x, ...)
}
\arguments{
\item{x}{An \code{PremiumFit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{PremiumFit}
}
