% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{print.RatioFit}
\alias{print.RatioFit}
\title{Print an \code{RatioFit} object}
\usage{
\method{print}{RatioFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"RatioFit"}.}

\item{...}{Unused.}
}
\description{
Print an \code{RatioFit} object
}
