% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{print.SAFit}
\alias{print.SAFit}
\title{Print method for \code{SAFit}}
\usage{
\method{print}{SAFit}(x, ...)
}
\arguments{
\item{x}{A \code{SAFit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{SAFit}
}
