% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capecod.R
\name{summary.CapeCodFit}
\alias{summary.CapeCodFit}
\title{Summary method for \code{CapeCodFit}}
\usage{
\method{summary}{CapeCodFit}(object, ...)
}
\arguments{
\item{object}{A \code{CapeCodFit} object.}

\item{...}{Unused.}
}
\description{
Summary method for \code{CapeCodFit}
}
