% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{summary.CCFit}
\alias{summary.CCFit}
\title{Summary method for \code{CCFit}}
\usage{
\method{summary}{CCFit}(object, ...)
}
\arguments{
\item{object}{A \code{CCFit} object.}

\item{...}{Unused.}
}
\description{
Summary method for \code{CCFit}
}
