% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{summary.ExposureFit}
\alias{summary.ExposureFit}
\title{Summary method for \code{ExposureFit}}
\usage{
\method{summary}{ExposureFit}(object, ...)
}
\arguments{
\item{object}{An \code{ExposureFit} object.}

\item{...}{Unused.}
}
\description{
Summary method for \code{ExposureFit}
}
