% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prem.R
\name{summary.PremFit}
\alias{summary.PremFit}
\title{Summary method for \code{PremFit}}
\usage{
\method{summary}{PremFit}(object, ...)
}
\arguments{
\item{object}{A \code{PremFit} object.}

\item{...}{Unused.}
}
\description{
Summary method for \code{PremFit}
}
