% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premium.R
\name{summary.PremiumFit}
\alias{summary.PremiumFit}
\title{Summary method for \code{PremiumFit}}
\usage{
\method{summary}{PremiumFit}(object, ...)
}
\arguments{
\item{object}{A \code{PremiumFit} object.}

\item{...}{Unused.}
}
\description{
Summary method for \code{PremiumFit}
}
