% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{summary.RatioFit}
\alias{summary.RatioFit}
\title{Summary method for \code{RatioFit}}
\usage{
\method{summary}{RatioFit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"RatioFit"}.}

\item{...}{Unused.}
}
\value{
A \code{data.table} with one row per cohort.
}
\description{
Summary method for \code{RatioFit}
}
