% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{summary.SAFit}
\alias{summary.SAFit}
\title{Summary method for \code{SAFit}}
\usage{
\method{summary}{SAFit}(object, ...)
}
\arguments{
\item{object}{A \code{SAFit} object.}

\item{...}{Unused.}
}
\description{
Per-cohort ultimate loss, SE, and CV.
}
