% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\name{.detect_maturity}
\alias{.detect_maturity}
\title{Internal: locate the first mature ata link from an \code{ATASummary}}
\usage{
.detect_maturity(
  x,
  max_cv = 0.15,
  max_rse = 0.05,
  min_valid_ratio = 0.5,
  min_n_valid = 3L,
  min_run = 2L
)
}
\description{
Internal: locate the first mature ata link from an \code{ATASummary}
}
\keyword{internal}
