% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{.expand_grid}
\alias{.expand_grid}
\title{Expand a \code{Triangle} object to a full projection grid}
\usage{
.expand_grid(triangle, ed_fit, prem_ata_fit, target, exposure)
}
\description{
Expand a \code{Triangle} object to a full projection grid
}
\keyword{internal}
