% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{.sa_param_var}
\alias{.sa_param_var}
\title{Hybrid parameter variance for a single cohort}
\usage{
.sa_param_var(
  loss_proj,
  exposure_proj,
  g_var,
  f_var,
  f_sel,
  last_obs,
  maturity_from,
  method = "sa"
)
}
\description{
Internal helper for parameter variance:

\itemize{
\item ED phase:
\eqn{\text{param}_{k+1} = \text{param}_k
      + (C^P_k)^2 \cdot \mathrm{Var}(\hat{g}_k)}
\item CL phase:
\eqn{\text{param}_{k+1} = f_k^2 \cdot \text{param}_k
      + (C^L_k)^2 \cdot \mathrm{Var}(\hat{f}_k)}
}
}
\keyword{internal}
